site/_databricks/requirements.txt
databricks-connect>=18.0.0
python-dotenv
ipykernel
